Revert Gemma4 to alpha's standalone pattern + rename upgrade.md → v2-to-v3-migration.md
Gemma4 fix:
- The Gemma 4 E2B/E4B smoke test caught a load-time keyNotFound error on
  every kv mode (`language_model.model.norm.weight` missing). Root cause:
  main's Gemma4Model wraps Gemma4TextModel via @ModuleInfo(key:
  "language_model"), expecting a wrapper-style sanitize that prepends
  `language_model.model.` — but alpha's Gemma4TextModel was designed as a
  top-level model whose sanitize *strips* the language_model. prefix.
  Combining the two paths made keys mismatch in both directions.
- Reverted to alpha's standalone pattern: register both "gemma4" and
  "gemma4_text" → Gemma4TextConfiguration / Gemma4TextModel.
- Removed Libraries/MLXLLM/Models/Gemma4Text.swift (was a copy of alpha's
  Gemma4.swift content with `vocabularySize` made mutable for the shim).
- Restored Libraries/MLXLLM/Models/Gemma4.swift to alpha's full 1207-line
  text impl (with all alpha's perf work intact: fused gate_up_proj,
  compile()-cached SharedMLP, symbolic .slidingWindow SDPA mask, MoE
  support for 26B-A4B).
- Confirmed: Gemma 4 E2B 4bit summarization passes 3/3 KV modes (none,
  affine4, turbo4v2) at ctx 128/1024/4096/32768 with coherent output.

Doc rename per user direction (keep our naming convention):
- `upgrade.md` → `v2-to-v3-migration.md`. Cross-references updated in
  Documentation.md and using.md. README link + URL anchors swapped.

README structure (user edits preserved):
- New `## Upgrading` parent section with `### Migrating to Version 3` as a
  subsection — gives a clean place to add `### Migrating to Version 4`
  later without churn.
- "Multiple KV compression algorithms" bullet added to the key features.
